requirements.txt
asgiref==3.7.2
Django==4.0.1
sqlparse==0.4.4
tzdata==2023.3
psycopg2-binary==2.9.6
djangorestframework~=3.13.0